+++ RELEASE_NOTES.md
@@ -33,6 +33,7 @@ Some things that are planned for the 7.0 release:
 
 ### 🐛 Bug fixes
 
+* Some system keyboard character margin actions have been corrected.
 * Standard English now also shows emoji skin tone variants when using KeyboardKit Pro.